delete unused struct
